refoctored pom.xml by setting all versions in properties, refactored methouds for geting random doctor and patient ids from db by getting id only, not all object. Moved class PatientResponse to patient directory
