+++ src/main/java/lviv/syrovyi/health_care/patient/service/PatientService.java
@@ -1,7 +1,6 @@
 package lviv.syrovyi.health_care.patient.service;
 
-import lviv.syrovyi.health_care.common.dto.response.PageResponse;
-import lviv.syrovyi.health_care.common.dto.response.PatientResponse;
+import lviv.syrovyi.health_care.patient.controller.dto.response.PatientResponse;
 import lviv.syrovyi.health_care.patient.controller.dto.request.PatientRequestDTO;
 import lviv.syrovyi.health_care.patient.controller.dto.response.PatientResponseDTO;
 import lviv.syrovyi.health_care.patient.controller.dto.response.PatientVisitsResponseDTO;
