+++ src/main/resources/application-dev.yml
@@ -4,3 +4,5 @@ spring:
     username: root
     password: root
     driver-class-name: com.mysql.cj.jdbc.Driver
+  jpa:
+    show-sql: true
