+++ src/main/resources/application.yml
@@ -2,13 +2,10 @@ server:
   port: 8080
 spring:
   profiles:
-    active: dev
+    active: docker
   application:
     name: "healthCare"
   jpa:
     properties:
       hibernate.jdbc.time_zone: UTC
     database-platform: org.hibernate.dialect.MySQLDialect
-    hibernate:
-      ddl-auto: update
-    show-sql: true
